Initial support for inheritance of custom casting
If the parent class is set up for custom casting, child classes are too, and
can either inherit the parent implementation of `__doCast()` or define their
own. They are also free to define an override for the inherited mtehod.
Furthermore, they can have the attribute themselves and/or implement the
interface directly without errors.

At the moment, if the parent class is not set up for custom casting, child
classes' desire for custom casting is ignored, since the point at which the
attributes are validated and we change the object handlers comes before the
point when classes with parent classes have their own object handlers overriden
with the parent's handlers.